Refactor common functions into a Python library
Refactor common functions into a centralized Python library.

* Move `sanitize_path`, `tfe_token`, `get_requests_session`, and `mod_dependencies` functions from `helpers.py` to `tfe_tools/common.py`.
* Update import statements in various files to import these functions from `tfe_tools/common.py`.
* Remove the moved functions from `helpers.py`.
* Add necessary imports in `tfe_tools/common.py` to support the moved functions.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/HappyPathway/tfe-tools?shareId=XXXX-XXXX-XXXX-XXXX).
