+++ utils/find_all_mods.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 import json
+from tfe_tools.common import sanitize_path, tfe_token, get_requests_session, mod_dependencies
 
 def main(mod_source, input_file):
     with open(input_file) as input:
